Casting image to float when blurring

+++ correct_jitter.py
@@ -30,7 +30,7 @@ class Jitter(object):
     def blurred_image(self):
         if self._blurred_image is None:
             print('Calculating new blurred image')
-            self._blurred_image = ndimage.gaussian_filter(self.image, self._blur_radius)
+            self._blurred_image = ndimage.gaussian_filter(self.image.astype(np.float64), self._blur_radius)
         return self._blurred_image
     
     @property
